refactor: unify data model and data access layer
- Create CorePlayer → RosteredPlayer → StatsPlayer → Player type hierarchy
- Add web/lib/data.ts as the single entry point for all Supabase data fetching
- Standardize on league_prices as the salary source of truth (remove transaction-override logic from player pages)
- Add web/lib/scoring.ts with calculateFantasyPoints() for Ottoneu Half PPR formula
- Deprecate Python analysis scripts (analyze_*.py) — TypeScript web UI is now the canonical calculation layer
- Update analysis.ts to delegate core fetching to data.ts
- Replace deprecated players.ts with re-exports from data.ts
- Update docs (ARCHITECTURE.md, CODE_ORGANIZATION.md)
