+++ docs/CODE_ORGANIZATION.md
@@ -6,8 +6,10 @@
 |------|------|---------|
 | Python config | `scripts/config.py` | All league constants, Supabase client factory |
 | TS config | `web/lib/config.ts` | Frontend constants (**must stay in sync with `scripts/config.py`**) |
-| TS types | `web/lib/types.ts` | All shared TypeScript interfaces |
-| Analysis math | `web/lib/analysis.ts` | TS port of `scripts/analysis_utils.py` |
+| TS types | `web/lib/types.ts` | All shared TypeScript interfaces (`CorePlayer → RosteredPlayer → StatsPlayer → Player`) |
+| Data layer | `web/lib/data.ts` | **Unified data access** — all Supabase fetching goes through here |
+| Scoring | `web/lib/scoring.ts` | Ottoneu Half PPR scoring formula (`calculateFantasyPoints`) |
+| Analysis math | `web/lib/analysis.ts` | Projection-enriched data + backtest fetching (builds on `data.ts`) |
 | Arb logic | `web/lib/arb-logic.ts` | Arbitration simulation logic |
 | DB schema | `schema.sql` | Canonical schema definition |
 | Migrations | `migrations/` | Numbered SQL migration files |
