+++ scripts/analyze_efficiency.py
@@ -1,3 +1,9 @@
+"""DEPRECATED: This script generates markdown reports that duplicate calculations now
+performed by the web UI (TypeScript). The web UI in web/lib/ is the canonical source
+of truth for VORP, surplus, arbitration, and projected salary calculations.
+This script is retained for reference only and may be removed in a future release.
+"""
+
 import os
 from typing import Optional
 import pandas as pd
