+++ scripts/analyze_projected_salary.py
@@ -1,4 +1,9 @@
-"""Salary analysis — keep vs. cut decisions for The Witchcraft roster.
+"""DEPRECATED: This script generates markdown reports that duplicate calculations now
+performed by the web UI (TypeScript). The web UI in web/lib/ is the canonical source
+of truth for VORP, surplus, arbitration, and projected salary calculations.
+This script is retained for reference only and may be removed in a future release.
+
+Salary analysis — keep vs. cut decisions for The Witchcraft roster.
 
 Uses surplus value (dollar_value - salary) which accounts for positional
 scarcity via VORP. This correctly values QBs higher in superflex.
